requirements.txt
fastapi>=0.100
uvicorn>=0.24
langchain>=0.1
langchain-huggingface>=0.1
langchain-ollama>=0.2
langchain-text-splitters>=0.2
pinecone-client>=3.0
pydantic>=2.0
pydantic-settings>=2.0
python-multipart>=0.0.6
pypdf>=3.0
pytesseract>=0.3
Pillow>=10.0
sentence-transformers>=2.0
httpx>=0.25
hypothesis>=6.0
pytest>=7.0
pytest-asyncio>=0.21
aiosqlite>=0.19
